+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Changelog
 
 ## _(in-progress)_
+* New: Extract code for creating link to category's archive into new `__c2c_linkify_categories_get_category_link()`
 * Change: Tweak descriptions to clarify that the links are to each category's archive
 * Change: Add default values for optional arguments to inline parameter documentation
 * Change: Note compatibility through WP 6.6+